Move the fake DNS logic in compatibility test, to avoid degrading the acme coverage
